cache class not actor
